+++ publisher/src/main/scala/org/podval/tools/publish/markup/DocBookXmlWriterConfig.scala
@@ -3,6 +3,7 @@ package org.podval.tools.publish.markup
 import org.podval.xml.XmlWriterConfig
 
 object DocBookXmlWriterConfig extends XmlWriterConfig(
+  selfCloseEmpty = true,
   preformat = Set("programlisting", "screen", "literallayout"),
   nest = Set("para", "simpara", "title"),
   cling = Set("footnote", "xref", "co")